+++ client/web/src/nav/NavLinks.tsx
@@ -87,7 +87,7 @@ const getMinimizableNavItems = (props: Props): JSX.Element[] => {
     const { showBatchChanges, settingsCascade } = props
 
     const settings = !isErrorLike(settingsCascade.final) ? settingsCascade.final : null
-    const { codeMonitoring } = settings?.experimentalFeatures || {}
+    const codeMonitoring = settings?.experimentalFeatures?.codeMonitoring !== false
     const codeInsights =
         settings?.experimentalFeatures?.codeInsights && settings?.['insights.displayLocation.insightsPage'] !== false